feat(2.0/dbscan): bf16 precision lock so distance dtype is not a lever
The points are handed out at bf16 precision (bf16-rounded values in an fp32
container, so the dbscan(x, eps, min_samples) contract is unchanged). The inputs
then carry only bf16 precision, so no solution can win by computing the eps-graph
pairwise distances in a lower dtype than the baseline -- bf16 tensor cores are the
floor, fp32/TF32 buy nothing, and anything below bf16 fails the gate. The blobs
are well separated (cluster_std=1.0, centers in +-10), so bf16 rounding never
flips an eps-neighbour: the reference's ARI stays exactly 1.0 on all four
workloads, no gate loosening needed.

speedup_target 51 -> 57 = 2x the flashlib reference geomean on the bf16 data
(28.6x on H100). images rebuilt to v0.3.0 (flash_gpu.py is baked at /opt).

+++ 2.0/problems/dbscan_gpu_kernel_optimization/flash_gpu.py
@@ -106,7 +106,12 @@ def _gpu_worker(payload: dict) -> dict:
             centers = (torch.rand(nc, D, generator=g, device=dev) * 2.0 - 1.0) * 10.0
             asg = torch.randint(0, nc, (N,), generator=g, device=dev)
             x = centers.index_select(0, asg) + torch.randn(N, D, generator=g, device=dev) * std
-            return {"x": x.to(torch.float32), "eps": float(w["eps"]),
+            # PRECISION LOCK (bf16): the points carry only bf16 precision, so a
+            # solution cannot win by computing the pairwise distances in a lower
+            # dtype than everyone else -- bf16 tensor cores are the floor, and the
+            # blobs are well separated (std=1.0, centers in +-10) so bf16 rounding
+            # never flips an eps-neighbour and the clustering is unchanged.
+            return {"x": x.to(torch.bfloat16).to(torch.float32), "eps": float(w["eps"]),
                     "min_samples": int(w["min_samples"])}
         x = torch.randn(w["N"], w["D"], generator=g, device=dev, dtype=torch.float32)
         if prim == "kmeans":
